refactor: s2 reconnect in an extra module
